Remove redundant error checking in process_course
Many error checking is redundant. The errors that the lines are
intending to catch should have already be caught by regex parsing.
For example, COURSE_REGEX requires the department regex group to
contain one or more capital english alphabets [A-Z]+. Therefore,
if department is empty or contain slash, ScrapeError will be raised
from line 234 since the regex does not match. Similarily, if
open_seats or total_seats is less than 0, the regex wouldn't be
matched and the error on line 278 will be raised.